engine update

+++ src/states/play.h
@@ -75,7 +75,7 @@ class _PlayState : public ae::_State {
 		void Render(double BlendFactor) override;
 
 		void GenerateHitEffects(_Entity *Attacker, const int Type, const _Hit &Hit, bool Death, float Rotation, bool CreateWallDecal=true);
-		void GenerateDamageText(const glm::vec2 &Position, int Value, bool Crit, bool HitPlayer);
+		void GenerateDamageText(const glm::vec2 &Position, int64_t Value, bool Crit, bool HitPlayer);
 		_Particle *GenerateTextParticle(const glm::vec2 &Position, const std::string &Value, const glm::vec4 &Color=glm::vec4(1.0f));
 		void GenerateExplosion(const _ParticleTemplate *ParticleTemplate, const glm::vec2 &Position, const glm::vec2 &Scale);
 		void GenerateProjectileEffects(const _ParticleTemplate *ParticleTemplate, const glm::vec2 &Position, const glm::vec4 &Color);
